Add DeepHit, SurvTRACE models and optimize codebase
