+++ pyproject.toml
@@ -46,11 +46,11 @@ Issues = "https://github.com/Cameron-Lyons/survival/issues"
 dev = [
   "pre-commit==4.5.1",
   "pytest==9.0.2",
-  "numpy==2.4.1",
+  "numpy==2.4.0",
 ]
 test = [
   "pytest==9.0.2",
-  "numpy==2.4.1",
+  "numpy==2.4.0",
   "pandas==2.3.3",
   "polars==1.37.1",
 ]
